Feature/docker persistencia

+++ src/main/kotlin/ar/edu/unsam/proyecto_de_sofware/event_nexus/dto/authDTO.kt
@@ -4,4 +4,4 @@ data class NewAccountRequest(val username:String, val password:String, val email
 
 data class LoginRequest(val username:String, val password:String)
 
-data class LoginResponse(val id:Int, val role:String)
+data class LoginResponse(val id:Long, val role:String)
